requirements.txt
wheel
Django==3.0.7
djangorestframework
conllu>=1.3.2
psycopg2
environs
dj-database-url
gunicorn
chardet==3.0.4
python-ldap==3.3.0
django-auth-ldap==2.2.0
drf_yasg==1.17.0
